Move API client to new folder
