+++ src/JobsInFinland.Api.Productizer/Client/JobsInFinlandApiClient.cs
@@ -1,6 +1,6 @@
 using JobsInFinland.Api.Infrastructure.CodeGen.Model;
 
-namespace JobsInFinland.Api.Productizer;
+namespace JobsInFinland.Api.Productizer.Client;
 
 internal class JobsInFinlandApiClient : IJobsInFinlandApiClient
 {
